requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.20.0
matplotlib>=3.4.0
scikit-learn>=0.24.0
tqdm>=4.60.0
captum>=0.4.0  # For explanation methods
pillow>=8.2.0
pandas>=1.3.0
medmnist>=2.0.0  # For ChestMNIST dataset
